Set InstanceRoleName as stack template output
We allow Buildkite agents to assume other IAM roles for certain
deployment/operational actions. The role to be assumed needs to include
the role ARN within its assume role policy.

It would be great to thread a dependency between the stack and the
assume policy document via the stack output.

We use the existing `AutoScalingGroupName` stack output to create/update
scheduled actions on the ASG and would like to use the role name to aid
in other resources creation.