render build/mappings.yml before building build/aws-stack.yml
The ultimate output of the build pipeline for this project is rendering
a full build/aws-stack.yml cloud formation template.

build/mappings.yml is required for that to work (see the `test -f`
within the target), so lets make it an explicit dependency.

This change means the recommended build command in the Readme
(`make build`) should work out of the box (assuming valid AWS
credentials available for building the AMIs).