make build: render the mappings.yml file before aws-stack.yml
rendering aws-stack.yml depends on the mappings.yml file existing. I
previously setup that dependency directly (see PR #706), however that
broke CI where we want to manually manage the mappings.yml file.

This seems like it might be a way to ensure `make build` works in our
development environment (as documented in the README) while also keeping
the CI pipeline working.

This is an alternative solution to PR #706