add a 10sec pause before we request instance termination
When the buildkite-agent exits cleanly (generally after an idle
timeout), we ask the ASG to terminate the instance. The termination can
happen *very* quickly, oftern within 2-3 seconds.

We use the legacy awslogs tool[1] to send logs to cloudwatch, and it
buffers new logs for 5 seconds (or 32Kb, whichever comes first).

This results in a race condition - often the final logs from the agent
as it shutdown may not make it to cloudwatch before the instance
shutsdown.

Long term we assume migrating from awslogs to the new cloudwatch agent
might improve this situation. However, we don't plan to attempt that
migration vor the v5 release and we'd like to improve the chances
customers will have access to a full agent log in clouddwatch. Adding a
10 sec sleep before requesting termination is slightly-sad, but
pragmatic option.

Note that this (sadly) won't resolve the race condition that prevents
other useful logs from making it to cloudwatch, like the result of
calling `aws autoscaling terminate-instance-in-auto-scaling-group` and
systemd during shutdown