bump vendored ECR plugin from v2.1.0 to 2.1.1
There's no significant changes in this bump (full diff at [1]). The only
change is support for an older version of bash that isn't included in
the elastic stack AMIs.

Still, I thought it was worth updating ahead of the v5.0.0 release. At
least next time we won't need to repeat the chore of checking if the
v2.1.1 release has relevant changes for the elastic stack.

[1] https://github.com/buildkite-plugins/ecr-buildkite-plugin/compare/v2.1.0...v2.1.1