restore the SpotPrice parameter, but deprecate it
In the next major release (v6?) we'd like to remove the SpotPrice
parameter - customers can use the OnDemandPercentage parameter to opt
into using spot instances with a price cap equal to the OnDemand price.

However, in the 5.x series we don't want breaking changes like removing
a param. This commit restores SpotPrice and keeps the same semantics:

* if it's set to 0 then we use OnDemandPercentage to decide if spot
  instances are used. OnDemandPercentage defaults to 100%, so stacks
  with a SpotPrice of 0 will continue to have 100% OnDeman instances
  when they upgrade
* if it's set to >0 then we ignore OnDemandPercentage and force the
  stack to be 100% spot.

When we're prepping for the v6 release we can remove the SpotPrice
parameter and use only the OnDemandPercentage parameter to decide if
spot instances are used.

I've also expands the SpotPrice parameter description. Partly to clarify
how it has always worked, partly to indicate it's deprecated.