bump lifecycled from 3.0.2 to 3.2.0
There have been two releases of lifecycled in the past 3 months - 3.1.0
and 3.2.0.

Neither has any changes that should impact the elastic stack, but it's
nice to stick near the latest release so the delta between the elastic
stack and upstream doesn't get too big.

Most of the changes are about updating the version of go we build with,
and updating some key dependencies (like aws-sdk-go)

[1] https://github.com/buildkite/lifecycled/releases/tag/v3.1.0
[2] https://github.com/buildkite/lifecycled/releases/tag/v3.2.0