bump docker from 19.03.x to 20.10.x
* Linux: 20.10.6
* Windows: 20.10.0

We're assuming this is a backwards compatible upgrade and safe to
include in the elastic stack 5.2.x to 5.3.x upgrade. Assuming the AMIs
build, we'll might merge and dogfood them for our internal builds to
build confidence in the compatibility and safeness of the bump.

The windows patch release is older because that's all that seems ot be
available.