allow multiple / characters in BuildkiteAgentTokenParameterStorePath
In #813 we added a regexp validation to this stack parameter that was
intended to check the parameter is either an empty string or a path that
starts with a / character.

The list of characters we permitted comes from [1]. However, we forgot
to include / as a permitted character beyond the first. It's valid (and
common) for paths to have multiple / to indicate hierarchy and we should
support it.

Fixes #834

[1] https://docs.aws.amazon.com/systems-manager/latest/userguide/sysman-paramstore-su-create.html