bk-install-elastic-stack.sh(cfn-env): Don't clobber user's env vars.
This change addresses the issue raised in
https://github.com/buildkite/elastic-ci-stack-for-aws/issues/891 - where a user sets an environment
variable in their pipeline config but it unexpectedly gets clobbered by Buildkite's defaults, which
are determined at AMI-build-time but reloaded before each build step.

We introduce a helper function, `set_unless_present`, which will raise a warning if a variable it
wants to set already exists, and avoid clobbering it.  Behaviour is unchanged for the "happy path".