hooks/environment: Create separate step for setting up env variables.
I initially tried to simply place the llama-line above the `~/cfn-env` source line.  Unfortunately
though, the llama-line depends on the contents of the `cfn-env` file for its version number!  But i
don't like lying console output, since in fact it has started setting up the environment before the
line saying as much has been printed.

My little compromise is to embrace the situation and introduce a new "environment variables" (Earth,
environment, get it?!  It has been a long lockdown, forgive me) step which may produce warnings and
will be expanded automatically if so.

I'm also removing the `cat ~/cfn-env` line, because that script is now a little smarter and will
print what it's doing, obviating the `cat`.  So, let's not confuse the user by outputting a bunch of
Bash script.