Don't accidentally send the token to logs on startup
The elastic stack sends its startup logs to the `/buildkite/elastic-stack/{instance_id}` log group in cloudwatch, which is basically a bash script with `set -x` enabled. As part of this process, we fetch the token from SSM Parameter store, but this is included in the output, meaning that the token gets leaked into cloudwatch logs, which are potentially much less privileged than the agent tokens themselves.

This PR disables set -x while we do secret things, and then turns it back on afterwards