Update description of BuildkiteAdditionalSudoPermissions parameter
The previous wording lead customers to believe that the command could be
specified with just its name. However, `sudoers(5)` requires that the
commands be specified with fully qualified paths. This lead some users
to believe that the stack was generating an invalid line in a sudoers
file when the error was the failure to fully qualify the path.