Remove goss timeouts
Most likely, commands were timing out because we ran some of the tests on `t3.nano`s, which have variable CPU, and the tests exhausted allowed CPU bursts. We conduct more intensive tests now, so we upgraded to `t3.small`. This may mean the longer timeouts are no longer needed.