Rename the parameter `EnableGitMirrorsExperiment` to `BuildkiteAgentEnableGitMirrors`
And no longer set the experiment in the agent config. Now that the
experiment has been mainlined, this feature is toggled by the non
emptiness of the `git-mirrors-path` config value. So all we need to do is
ensure that if the new parameter is true, it is set to a non-empty value
(and the path exists and is writable by the agent).

Fixes: #1121