Fix goss docker info commands
They need to have some level of nested templates because otherwise,
goss will attempt to evaluate the templates that are intended to be
evaluated by docker.

Also, the elements of `stdout` are regexes, so we
adapt it to test for each expected element of the list. Goss will report
which regex did not match the output, so we can use this determine which
plugin are missing. It won't be able to tell if there are other plugins,
but that is very much a feature.