Make `capacity-optimized` the default spot allocation strategy
Amazon's recommended strategy is `price-capacity-optimized`: https://docs.aws.amazon.com/autoscaling/ec2/APIReference/API_InstancesDistribution.html but the default strategy for a InstancesDistribution is `lowest-price`.
In the elastic ci stack v5, we overrode the default and set it to `capacity-optimized`.

However, for a CI system, I think the strategy with the lowest risk of interruption is preferable.
There are two distinct reasons for this:
- Interrupting a CI job has a cost in developer time. Usually, developers are
  more expensive than instances.
- The instances carry state, e.g. the checkout dir and docker caches.
  Loosing this state costs time on subsequent jobs, which means those
  instance could be running for longer, and developers could be waiting
  for longer.

This is clearly a tricky balancing act, so giving users the freedom to
select a different strategy is also a good idea.