Handle hard failures (eg. kernel panic) during bootstrap
This change configures cloud-init to run the bootstrap script on every
boot so that we can check for previous failed attempts and mark the
instance as unhealthy.

This caters for the scenario where the kernel panics during a bootstrap
attempt which leaves the instance non-operational on reboot but marked
healthy.