Update CHANGELOG for v6.3.0 again
I made a mistake in the previous commit. I forgot to remove the change
that updated the changelog with the changes in v5.22.3 (which is not on
the main branch) from the `Changed` section. It should be in the
`Internal` section, which was. In other words, the change was duplicated.