Remove redirecting echos and cats to stderr
The command that sends the scripts output streams to the loggers merges
stdout and stderr anyway, so there is no point in writing to stderr in
the script.

This is the logging command:
```
exec > >(tee -a /var/log/elastic-stack.log | logger -t user-data -s 2>/dev/console) 2>&1
```
it is present in each of these scripts.