Fix docker run does not recognize previously pulled image
When you pull an image from a registry and specify both a tag and a digest,
docker does not seem to tag it locally. This leaves the image unable to
be run if both the tag and the digest are specified to `docker run`.
But it will run if only by the digest is specified.