Restore SSM Session Manager Plugin
This was removed in the migration to Amazon Linux 3 with the belief that
there was no use case for the Session Manager Plugin being present on
the agent[1]. It has since been realised that it is quite useful. For
example, the agent may be used to start ECS Tasks which perform work in
other environments/network configurations (such as Database migrations
during deployments).

1. https://github.com/buildkite/elastic-ci-stack-for-aws/pull/1122#discussion_r1220593363