New pipeline: deregister old AMIs
We've been publishing AMIs to 20 regions for ~6 years, and we've built
quite a collection.

Many of them we want to keep indefinitely, like any that we've
published in a released Cloud Formation template on [1]. Or any that
were published in the last 12 months.

However, there's many many other images that it's safe to garbage
collect. For example, we publish AMIs to every region for every main
branch build, so we and customers can test them. They're useful for a
short time, but any older than 12 months we can remove. By that time,
customers have the option to move to a formal release with the features
they want.

[1] https://github.com/buildkite/elastic-ci-stack-for-aws/releases