Expand the AMI cleaning pipeline to all regions
All regions we publish the elastic stack to, anyway. There's 20 - any
region launched in the last few years and not requested by a customer
currently has no AMIs published to it.

The pipeline is still in dry run mode, so this will have no actual
impact on AMIs. The IAM role also lacks permission to actually delete
anything yet.

This will give logs of every region that I can examine and spot check
for correctness though.