Allow user to specify KMS key ARN for pipeline signing.
We have a use case where we want to have a single KMS key that Buildkite
agents in multiple accounts can use to sign and verify pipeline steps.
By letting the user specify the whole key ARN rather than just the key
ID, the user is free to use a key wherever they want.

* templates/aws-stack.yml: Rename PipelineSigningKMSKeyId to
PipelineSigningKMSKeyArn and avoid assuming the key is in the same
account/region as the agent stack.

Co-authored-by: Matty Boyles <mathew.boyles@redbubble.com>