Separate cleanup of CFn stacks & service roles
Sometimes our CloudFormation cleanup process deletes stacks in the wrong
order. A stack containing the service role for another stack can be
deleted before the stack that depends on it. This leaves the dependent
stack orphaned, requiring tedious manual intervention to fix and costing
us money in the meantime.

This change ensures that test stacks are deleted before their service
roles, which should prevent this race condition.