cleanup: Don't hold public AMIs sacred
We found that not every public AMI needs to be kept, and we are
retaining quite a few we don't need!

This change introduces an extra check before keeping a public AMI: it
checks the CloudFormation template(s) for the version(s) that the AMI is
tagged with, and ensures that the AMI is actually used by at least one
of those templates.

If it's not used, we fall through to the existing checks on age and last
launch date.